+++ src/Readarr.Api.V1/Users/UserWizardProgressResource.cs
@@ -1,0 +1,7 @@
+namespace Readarr.Api.V1.Users
+{
+    public class UserWizardProgressResource
+    {
+        public bool InProgress { get; set; }
+    }
+}